tests/wpt/tests/css/css-anchor-position/anchor-name-multicol-004.html

<!DOCTYPE html>
<title>Anchor name resolution of OOF anchors in multicol</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#propdef-anchor-name">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-size">
<link rel="author" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<style>
.relpos {
  position: relative;
}
.abspos {
  position: absolute;
}
.columns {
  column-count: 6;
  column-fill: auto;
  column-gap: 10px;
  column-width: 20px;
  width: 170px;
  height: 50px;
}
.spacer {
  height: 10px;
  background: pink;
}
.anchor {
  anchor-name: --a1;
  margin-left: 5px;
  width: 10px;
  background: orange;
}
.target {
  position: absolute;
  background: lime;
  opacity: .3;
  left: anchor(--a1 left);
  top: anchor(--a1 top);
  width: anchor-size(--a1 width);
  height: anchor-size(--a1 height);
}
</style>
<body onload="checkLayoutForAnchorPos('.target')">
  <!-- All targets should find the static positioned anchor -->
  <div class="spacer"></div>
  <div class="relpos">
    <div class="columns relpos">
      <div class="relpos">
        <div class="spacer"></div>
        <div class="anchor abspos" style="top: 120px; height: 100px"></div>
        <div class="anchor" style="height: 60px"></div>
        <div class="target"
             data-expected-width=40 data-expected-height=50></div>
      </div>
      <div class="target"
           data-expected-width=40 data-expected-height=50></div>
    </div>
  </div>
</body>
